+++ test-fixtures/plugins-pnpm/index.result.php
@@ -1,0 +1,14 @@
+<?php
+
+array_map(
+  function ($arg1, $arg2) use ($var1, $var2) {
+    return $arg1 + $arg2 / ($var + $var2);
+  },
+  [
+    "complex" => "code",
+    "with" => "inconsistent",
+    "formatting" => "is",
+    "hard" => "to",
+    "maintain" => true,
+  ]
+);